requirements.txt
python-telegram-bot==20.5
yt-dlp==2024.3.10